requirements.txt
langchain
langchain-groq
fastapi
uvicorn
python-dotenv
langsmith
ipykernel
pytest
pydantic


